feat(core): introduce lock_manager
it eats some functions from apps.base, and introduces an asynchronous
notification mechanism for suspending via `notify_suspend()`
