+++ core/embed/upymod/qstrdefsport.h
@@ -108,6 +108,7 @@ Q(apps.common.coins)
 Q(apps.common.definitions)
 Q(apps.common.definitions_constants)
 Q(apps.common.keychain)
+Q(apps.common.lock_manager)
 Q(apps.common.passphrase)
 Q(apps.common.paths)
 Q(apps.common.payment_request)
@@ -232,6 +233,7 @@ Q(keychain)
 Q(layout)
 Q(layouts)
 Q(list_names)
+Q(lock_manager)
 Q(log)
 Q(log_benchmark)
 Q(loop)
