+++ core/src/apps/base.py
@@ -7,14 +7,10 @@ from trezor.enums import HomescreenFormat, MessageType
 from trezor.messages import Success, UnlockPath
 from trezor.ui.layouts import confirm_action
 from trezor.wire import context
-from trezor.wire.message_handler import filters, remove_filter
+from trezor.wire.message_handler import filters
 
 from . import workflow_handlers
-
-if utils.USE_POWER_MANAGER:
-    from trezor import io
-    from trezor.power_management.autodim import autodim_display
-    from trezor.power_management.suspend import suspend_device
+from .common import lock_manager
 
 if TYPE_CHECKING:
     from typing import NoReturn
@@ -32,7 +28,6 @@ if TYPE_CHECKING:
         Ping,
         SetBusy,
     )
-    from trezor.wire import Handler, Msg
 
     if utils.USE_THP:
         from trezor.messages import (
@@ -273,7 +268,7 @@ if utils.USE_THP:
             channel_ctx=channel, session_id=session_id
         )
         try:
-            await unlock_device()
+            await lock_manager.unlock_device()
             await derive_and_store_roots(new_session, message)
         except DataError as e:
             return Failure(code=FailureType.DataError, message=e.message)
@@ -407,7 +402,7 @@ async def handle_Cancel(msg: Cancel) -> NoReturn:
 
 
 async def handle_LockDevice(msg: LockDevice) -> Success:
-    lock_device()
+    lock_manager.lock_device()
     return Success()
 
 
@@ -422,7 +417,7 @@ async def handle_SetBusy(msg: SetBusy) -> Success:
         context.cache_set_int(APP_COMMON_BUSY_DEADLINE_MS, deadline)
     else:
         context.cache_delete(APP_COMMON_BUSY_DEADLINE_MS)
-    set_homescreen()
+    lock_manager.set_homescreen()
     workflow.close_others()
     return Success()
 
@@ -521,137 +516,6 @@ async def handle_CancelAuthorization(msg: CancelAuthorization) -> protobuf.Messa
     return Success(message="Authorization cancelled")
 
 
-def set_homescreen() -> None:
-    import storage.recovery as storage_recovery
-
-    from apps.common import backup
-
-    set_default = workflow.set_default  # local_cache_attribute
-
-    if context.cache_is_set(APP_COMMON_BUSY_DEADLINE_MS):
-        from apps.homescreen import busyscreen
-
-        set_default(busyscreen)
-
-    elif not config.is_unlocked():
-        from apps.homescreen import lockscreen
-
-        set_default(lockscreen)
-
-    elif _SCREENSAVER_IS_ON:
-        from apps.homescreen import screensaver
-
-        set_default(screensaver, restart=True)
-
-    elif storage_recovery.is_in_progress() or backup.repeated_backup_enabled():
-        from apps.management.recovery_device.homescreen import recovery_homescreen
-
-        set_default(recovery_homescreen)
-
-    else:
-        from apps.homescreen import homescreen
-
-        set_default(homescreen)
-
-
-def lock_device(interrupt_workflow: bool = True) -> None:
-    if config.has_pin():
-        config.lock()
-        filters.append(_pinlock_filter)
-        set_homescreen()
-        if interrupt_workflow:
-            workflow.close_others()
-
-
-def lock_device_if_unlocked() -> None:
-    from apps.common.request_pin import can_lock_device
-
-    if not utils.USE_BACKLIGHT and not can_lock_device():
-        # on OLED devices without PIN, trigger screensaver
-        global _SCREENSAVER_IS_ON
-
-        _SCREENSAVER_IS_ON = True
-        set_homescreen()
-
-    elif config.is_unlocked():
-        lock_device(interrupt_workflow=workflow.autolock_interrupts_workflow)
-
-    if utils.USE_POWER_MANAGER and not utils.EMULATOR:
-        # FIXME: suspend not implemented on emulator
-        wakeup_flag = suspend_device()
-        handle_wakeup_from_suspend(wakeup_flag)
-
-
-if utils.USE_POWER_MANAGER:
-
-    def lock_device_if_unlocked_on_battery() -> None:
-        """Lock the device if it is unlocked and running on battery or wireless charger."""
-        if not io.pm.is_usb_connected():
-            lock_device_if_unlocked()
-
-    def handle_wakeup_from_suspend(wakeup_flag: int) -> None:
-        """Handle wakeup from suspend."""
-        from trezor.ui import CURRENT_LAYOUT
-
-        if wakeup_flag == io.pm.WAKEUP_FLAG_BUTTON:
-            if CURRENT_LAYOUT is not None:
-                CURRENT_LAYOUT.layout.request_complete_repaint()
-        # TODO: handle PWR
-
-
-async def unlock_device() -> None:
-    """Ensure the device is in unlocked state.
-
-    If the storage is locked, attempt to unlock it. Reset the homescreen and the wire
-    handler.
-    """
-    from apps.common.request_pin import verify_user_pin
-
-    global _SCREENSAVER_IS_ON
-
-    if not config.is_unlocked():
-        # verify_user_pin will raise if the PIN was invalid
-        await verify_user_pin()
-
-    _SCREENSAVER_IS_ON = False
-    set_homescreen()
-    remove_filter(_pinlock_filter)
-
-
-def _pinlock_filter(msg_type: int, prev_handler: Handler[Msg]) -> Handler[Msg]:
-    if msg_type in workflow.ALLOW_WHILE_LOCKED:
-        return prev_handler
-
-    async def wrapper(msg: Msg) -> protobuf.MessageType:
-        await unlock_device()
-        return await prev_handler(msg)
-
-    return wrapper
-
-
-# this function is also called when handling ApplySettings
-def reload_settings_from_storage() -> None:
-    from trezor import ui
-
-    workflow.idle_timer.set(
-        storage_device.get_autolock_delay_ms(), lock_device_if_unlocked
-    )
-
-    if utils.USE_POWER_MANAGER:
-        # autodim setting is not from storage but keeping it here for simplicity
-        workflow.idle_timer.set(30_000, autodim_display)
-        workflow.idle_timer.set(
-            storage_device.get_autolock_delay_battery_ms(),
-            lock_device_if_unlocked_on_battery,
-        )
-    wire.message_handler.EXPERIMENTAL_ENABLED = (
-        storage_device.get_experimental_features()
-    )
-    if ui.display.orientation() != storage_device.get_rotation():
-        ui.backlight_fade(ui.BacklightLevels.DIM)
-        ui.display.orientation(storage_device.get_rotation())
-
-
 def boot() -> None:
     from apps.common import backup
 
@@ -676,9 +540,10 @@ def boot() -> None:
     ]:
         workflow_handlers.register(msg_type, handler)
 
-    reload_settings_from_storage()
+    lock_manager.reload_settings_from_storage()
     if backup.repeated_backup_enabled():
         backup.activate_repeated_backup()
     if not config.is_unlocked():
         # pinlocked handler should always be the last one
-        filters.append(_pinlock_filter)
+        # TODO: hide this inside lock_manager
+        filters.append(lock_manager._pinlock_filter)
