+++ core/src/apps/webauthn/fido2.py
@@ -11,8 +11,8 @@ from trezor.crypto.curve import nist256p1
 from trezor.ui import Layout
 from trezor.ui.layouts import error_popup
 
-from apps.base import set_homescreen
 from apps.common import cbor
+from apps.common.lock_manager import set_homescreen
 
 from . import common
 from .credential import Credential, Fido2Credential
